Sha256 Rainbow Table Crack
This is a rainbow table generation and crack of SHA256. It can easily
be modified to different reduction functions. Currently the code is
very rough and will be improved in a while. With the current version a
rainbow table of 100 mb would give > 80% password cracking for 6
character lowercase alphabetic passwords.